+++ src_new/aliases/levi_ataxia/for_levi/levi_062424/configs/015_Gem_Cloaking.lua
@@ -1,0 +1,24 @@
+--[[mudlet
+type: alias
+name: Gem Cloaking
+hierarchy:
+- Levi_Ataxia
+- Ataxia
+- Basher
+- Configs
+attributes:
+  isActive: 'yes'
+  isFolder: 'no'
+regex: ^abgcuse$
+command: ''
+packageName: ''
+]]--
+
+if not ataxiaBasher.gemCloaking then
+	ataxiaBasher.gemCloaking = true
+	ataxiaEcho("Enabled gem of cloaking auto-activation in Moghedu.")
+else
+	ataxiaBasher.gemCloaking = false
+	ataxiaEcho("Disabled gem of cloaking auto-activation in Moghedu.")
+end
+ataxia_saveSettings(false)